Adding support for Raspberry Pi1
The build script build_bal.sh now also supports building the module for raspberry pi1
For that purpose, the buildscript now needs to call with one input parameter:

  ./build_bal.sh raspi1

builds for raspberry pi 1

  ./build_bal.sh raspi2

builds for raspi2

Note: So far the script does everything, i.e. it compiles the module, the dt overlay file and
installs it directly within /lib/modules and /boot, repectively. So the script should be executed on the target machine.